Fix #48, no input passed to table tool
Return an error code (EXIT_FAILURE) if no input files were passed to the
eds2cfetbl tool.  This points to an error in the parent script/makefile
that should be addressed and thus the build should stop.  With a success
return, the build keeps running as if everything is OK but there will be
no table files at the end.