Fix #85, updates to struct/typedef names
Implements a bit of paradigm shift in how names of symbols/types are
generated, relying on prefixes rather than suffixes.  Using suffixes
has an opportunity of name collisions that is not possible when using
prefixes.

A separate typedef file maps to the existing type names and thus keeps
it compatible with existing code, but only if that typedef file is used.